dev_test/pyproject.toml
[project]
name = "dev-test"
version = "0.1.0"
description = ""
authors = [
    {name = "luis-barbara",email = "luis.carlos.barbara@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.95.0",
    "openai>=0.27.0",
    "python-dotenv>=0.21.0",
    "uvicorn>=0.22.0",
    "transformers (>=4.49.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "dev_test", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
